MuDataDataset: error out early if mudata.axis != 0
MuData objects with axis=1 can violate our fundamendal data model
of local and global alignment: Since group assignments are taken
from the global .obs, the same group may be present in multiple
AnnData objects with only partially overlapping features. In this
case it is not clear what the local samples and features should be.