Refactor the priors (#153)
* begin refactoring priors to make them fully modular

- introduce a new Prior base class and subclasses Horseshoe and
  GaussianProcess. These wrap the Pyro priors, similar to how
  likelihoods work
- The Horseshoe and GaussianProcess classes are now responsible for
  extracting their covariates from the data and setting up their factors

Training works, post-fit processing does not work yet

* move all of GP handling inside the GP prior

it's now responsible for initialization and time warping

* disable pcgse for now

* make each prior responsible for saving and loading its state

* first stab at dynamic API pickup from priors

currently only annotations from the Horseshoe prior work

* make the Horseshoe prior responsible for PCGSE

* register Horshoe API only on HS instances that have annotations

* make GP prior responsible for GPs

It now stores the fitted GPs and calculates GP values for new x values.

note: tab completion for parameters in Jupyter Lab only works with
%config IPCompleter.use_jedi = False
I don't know enough about jedi at the moment to figure out how to fix
this

* add support for dynamic properties

the HS prior is now responsible for n_informed_factors

* make the GP prior responsible for all its properties (covariates etc)

* move handling of SnS-specific attributes/methods out of MOFAFLEX class

* fix dynamic method name conflict resolution

Let each prior specify if and where to insert the axis name (factor or
weight) into the dynamic API, instead of attempting automatich conflict
handling. This makes the names more predictable and suitable for
downstream APIs, e.g. plotting

* remove last vestiges of return type selection for result getters

Always return dataframes (except for factors, where we use the AnnData
return type to also return covariates). It's trivial to get a numpy
array or an AnnData object from a dataframe and it simplifies the
prior-specific dynamic API handling a lot.

The AnnData return type in get_factors is currenlty used by our plotting
functions. TODO: Think about explicitly passing in the data together
with the model to the plotting functions. That would eliminate the need
for us to store and return obs metadata, and it would fix the case where
metadata that is added to the data after the model is trained cannot be
plotted with our plotting functions.

* covariates: use dataframes throughout

- datasets return covariates as dataframes
- priors store them as dataframes

this makes for better handling of categoricals in guiding factors and
less boilerplate for dynamic API getters (no need to pass in
factor_names and nonfactor_names) every time

* allow dynamic API properties to be ordered by factors

* simplify

* store results directly as DataFrames

this simplifies the getters (no need to convert to DF) and the
on_train_end hook, since there is no need to pass factor_names and
nonfactor_names as arguments anymore

* GP: fix get_gps() where x has different n of samples than trained on

* rename Horseshoe to InformedHorseshoe

this avoids having to define the user-facing API dynamically after
training

* simplify Prior._api decorator

only needs to work as a decorator now, not as a method

* set up dynamic API at import time

- resolve name conflicts at import time
- generate dummy methods/properties at import time with the correct
  documentation, so Sphinx can pick them up

docstring parsing/processing is still a tad brittle

* datasets: fix dataframe dtype conversion

* simplify

* fix loading saved models

* invert logic of how known priors are detected

previously, PyroPriors were prioritized and had to contain _factors and
_weights attributes. Now, wrapper Priors are prioritized and must
contain _factors and _weights attributes, PyroPriors are assumed to be
suitable for both factors and weights by default. The rationale being
that anything simple enough to not need a wrapper class is very likely
suitable for both factors and weights.

* add docstrings to Prior base class

* fixup! covariates: use dataframes throughout

* fix MOFA compat storage, device handling during training

* go back to storing numpy arrays instead of DataFrames for most things

This partially reverts ef0ddb6. Turns out storing dataframes has some
issues:
- Dramatically increased file size for the saved model. This is because
  weights and covariates for weight priors are stored as factors x
  weights. When saving to HDF5, each column is stored a separate array.
  These arrays are quite short (n_factors), and the overhead appears to be
  quite large.
- Column name and DF indices must be strings
- Column names cannot contain slashes (this is an issue when deriving
  factor names from an annotation, as sometimes there are slashes in
  annotation names)
- For real-world data with many features, storing factors x weights
  dataframes is impossible: AnnData writes the order of the columns as a
 string array into the HDF5 group attribute for the DF, and the attribute
 is limited to 69 KiB. Trying to save a model trained on a large dataset
 raises an exception.

* make tests work again

* add test for Prior base class

* store all model results and covariates with factors in axis 1

this requires less special casing in the priors and dynamic API wrappers