rework the priors API (#156)
* generate user-facing wrapper classes for priors

Also restructure documentation. Now that we have a bit more material for
the main API page, we no longer need automodsumm and can use the
standard autosummary. Unfortunately, we still need  a custom module
template to make autosummary automatically generate stub pages for the
module contents.

* remove presets for now (to be re-added after the new API is finalized)

* the MOFAFLEX class now accepts instances of the Prior API classes

* allow sequences as keys in ModelOptions.factor_prior and weight_prior

If e.g. a GP prior with many parameters is used for a subset of groups,
it can now be specified only once, with the groups given as a tuple key
of the dict, instead of having to create multiple identical
GaussianProcess objects.

* fix tests

* raise exception if a prior is used on the wrong axis